requirements.txt
litellm==1.44.7
python-dotenv==1.0.1
rich==13.8.0
